requirements.txt
requests==2.25.1
win10toast==0.9
pync==2.0.3
Pillow==8.3.2
PyQt5==5.15.4
